requirements.txt
matplotlib
nltk
numpy
keras
tensorflow
pydot
graphviz
flask
pyspellchecker